PLAT-98205: Separate logics in moonstone/VirtualListBase, ScrollerBase, and ScrollableBase(Native) into custom Hooks (#2671)
* Divide moonstone/VirtualListBase

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Cleanup

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Divide moonstone/Scroller

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix indentation

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Divide moonstone/Scrollable

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix indendation

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Divide moonstone/Scrollable.useSpottable

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Refactoring

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Divide ScrollableNative

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix issues mentioned

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Cleanup

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Cleanup

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Cleanup

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Cleanup

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Rename * useEffects as * Hooks

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Revise key event handlers

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Add useRef import

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Refresh from scroller-decouple

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix lint

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix lint

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Update Scroller-specs.js

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Update Scroller-specs.js

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix issues

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)

* Fix lint

Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)
