+++ packages/moonstone/Scrollable/useEventTouch.js
@@ -1,0 +1,34 @@
+import Spotlight from '@enact/spotlight';
+
+const useEventTouch = (props, instnaces, dependencies) => {
+	/*
+	 * Dependencies
+	 */
+
+	const {isScrollButtonFocused} = dependencies;
+
+	/*
+	 * Functions
+	 */
+
+	function handleTouchStart () {
+		const focusedItem = Spotlight.getCurrent();
+
+		if (!Spotlight.isPaused() && focusedItem && !isScrollButtonFocused()) {
+			focusedItem.blur();
+		}
+	}
+
+	/*
+	 * Return
+	 */
+
+	return {
+		handleTouchStart
+	};
+};
+
+export default useEventTouch;
+export {
+	useEventTouch
+};
